test(ios): compile the two new runner assertions
Both were added without compiling the runner target, which is not reachable from
the differential lane: `SnapshotRect` takes Doubles and the new fold test spelled
its y offset in Ints, and the fold test reached for `RawAXNode.replacing`, which
is internal to the presentation module and invisible to the runner tests. The
fold cases now state their source bit where the node is built, through a defaulted
parameter on the test's own builder, which is also what `normalized()` leaves
behind when the viewport read failed.

Verified with the runner test target built with
`AGENT_DEVICE_XCUITEST_INCLUDE_UNIT_TESTS=1` and run on the `rnav-repro`
simulator: 9 selected cases, 0 failures, including both new ones and
`testRegularFoldKeepsWindowCarriersButNeverHittableOutsideClip`, which is the
reported-viewport neighbour of the new fold policy.

+++ apple/runner/AgentDeviceRunner/AgentDeviceRunnerUITests/UnitTests/RunnerTests+SnapshotPresentationTests.swift
@@ -635,7 +635,7 @@ extension RunnerTests {
     ) -> RawAXNode {
       RawAXNode(
         index: index, type: type, label: label, identifier: nil, value: nil,
-        rect: SnapshotRect(x: 10, y: 20 + index * 60, width: 100, height: 44),
+        rect: SnapshotRect(x: 10, y: Double(20 + index * 60), width: 100, height: 44),
         enabled: enabled, focused: nil, selected: nil, hittable: false,
         depth: depth, parentIndex: parent, hiddenContentAbove: nil, hiddenContentBelow: nil
       )
